mnl - mjson_str2hdf() bug fix
Former-commit-id: 555a326b57dd03a86776954680d64e43e5cf2b4f

+++ lib/mnl/mjson.c
@@ -110,7 +110,7 @@ void mjson_str2hdf(HDF *node, struct json_object *o)
     if (!obj && s && *s) {
         obj = json_tokener_parse(s);
     }
-    if (!obj || obj < 0) return;
+    if (!obj || (int)obj < 0) return;
 
     type = json_object_get_type(obj);
